add norm
